// types/leda-type.ts
export type Leads = {
  id: string
  location: string
  leadFrom: string
  externalId: string
  lattitude: string
  longitude: string
  date: string
  address: string
  city: string
  state: string
  scope: string
  zip: string
  firstName: string
  lastName: string
  customerAddress: string
  email: string
  leadsAge: string
  phone: string
  checkboxAddress: boolean
  customerState: string
  customerCity: string
  customerZip: string
}
export type UserAddress = {
  city: string
  state: string
  zip: string
  fullAddress: string
}